refactor: jsx파일 tsx 파일 형식으로 변경
